small refactoring Eip-1559 related functionality
